Fix bug for gradle build
Currently gradle is unable to build jar file due to missing
javafx but javafx is already included in the dependency.

The bug is due to the main file extending application. Add a wrapper
class file to launch the application.
