+++ .gitignore
@@ -17,4 +17,3 @@ bin/
 text-ui-test/EXPECTED-UNIX.TXT
 
 /data/duke.txt
-/data/tasks.txt
