+++ src/main/java/duke/Duke.java
@@ -71,7 +71,6 @@ public class Duke {
     public String getResponse(String input) {
         try {
             Command command = Parser.parse(input);
-            System.out.println(command);
             return command.execute(tasks, ui, storage);
         } catch (DukeException e) {
             return e.getMessage();
