+++ src/main/java/duke/Launcher.java
@@ -1,0 +1,38 @@
+package duke;
+
+import java.io.IOException;
+
+import duke.controller.MainWindow;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+
+
+/**
+ * duke.Main entry point to run the java app
+ */
+public class Launcher extends Application {
+    private static final String FILE_PATH = "data/duke.txt";
+
+    private final Duke duke = new Duke(FILE_PATH);
+    @Override
+    public void start(Stage stage) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(Launcher.class.getResource("/view/MainWindow.fxml"));
+            AnchorPane ap = fxmlLoader.load();
+            Scene scene = new Scene(ap);
+            stage.setScene(scene);
+            stage.setTitle("ANYACHAT");
+            fxmlLoader.<MainWindow>getController().setDuke(duke);
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void launch() {
+        Application.launch();
+    }
+}
