+++ src/main/java/duke/Main.java
@@ -1,34 +1,13 @@
 package duke;
 
-import java.io.IOException;
-
-import duke.controller.MainWindow;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
-
 /**
- * Main entry point to run the java app
+ *
  */
-public class Main extends Application {
-    private static final String FILE_PATH = "data/duke.txt";
-
-    private final Duke duke = new Duke(FILE_PATH);
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            stage.setTitle("ANYACHAT");
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+public class Main {
+    /**
+     * @param args
+     */
+    public static void main(String[] args) {
+        new Launcher().launch();
     }
 }
