+++ src/main/java/duke/Parser.java
@@ -55,15 +55,12 @@ public class Parser {
             desc = fullCommand.substring(TLENGTH);
             return new Todo(desc);
         case EVENT_SYNTAX:
-            System.out.println("debugging event: " + fullCommand);
             commandParts = fullCommand.substring(ELENGTH).split(" /at ");
-            System.out.println(Arrays.toString(commandParts));
             desc = commandParts[0];
             time = commandParts[1];
             return new Event(desc, time);
         case DEADLINE_SYNTAX:
             commandParts = fullCommand.substring(DLENGTH).split(" /by ");
-            System.out.println(Arrays.toString(commandParts));
             desc = commandParts[0];
             time = commandParts[1];
             return new Deadline(desc, time);
